Fix Avalonia context menu to appear next to mouse pointer rather than along the top or bottom of the control, and demonstrate usage in the "Custom Right-Click Menu" control demo.
This implementation uses reflection to access a private 3-argument Open() method in Avalonia rather than the standard public 1-argument version. The 3-argument version allows us to specify that the context menu was requested by the mouse (rather than, say, the Menu key or Shift-F10) and should be placed next to the pointer.